feat: support paragraph publishable
