+++ zeppelin-web-angular/src/app/core/paragraph-base/public-api.ts
@@ -10,24 +10,5 @@
  * limitations under the License.
  */
 
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { NotebookComponent } from './notebook.component';
-
-const routes: Routes = [
-  {
-    path: ':noteId',
-    component: NotebookComponent
-  },
-  {
-    path: ':noteId/revision/:revisionId',
-    component: NotebookComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class NotebookRoutingModule {}
+export * from './paragraph-base';
+export * from './published';
