+++ zeppelin-web-angular/src/app/core/paragraph-base/published.ts
@@ -10,24 +10,8 @@
  * limitations under the License.
  */
 
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+export const publishedSymbol = Symbol('published');
 
-import { NotebookComponent } from './notebook.component';
-
-const routes: Routes = [
-  {
-    path: ':noteId',
-    component: NotebookComponent
-  },
-  {
-    path: ':noteId/revision/:revisionId',
-    component: NotebookComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class NotebookRoutingModule {}
+export interface Published {
+  readonly [publishedSymbol]: true;
+}
