+++ zeppelin-web-angular/src/app/core/public-api.ts
@@ -13,3 +13,4 @@
 export * from './message-listener';
 export * from './destroy-hook';
 export * from './copy-text';
+export * from './paragraph-base';
