+++ zeppelin-web-angular/src/app/pages/workspace/notebook/paragraph/paragraph.component.html
@@ -48,6 +48,7 @@
                                          (editorHideChange)="commitParagraph()"
                                          (enabledChange)="commitParagraph()"
                                          (titleChange)="commitParagraph()"
+                                         (openSingleParagraph)="openSingleParagraph($event)"
                                          (runOnSelectionChangeChange)="commitParagraph()"
                                          (runParagraph)="runParagraph()"
                                          (moveUp)="moveUpParagraph()"
