+++ zeppelin-web-angular/src/app/pages/workspace/published/paragraph/paragraph.component.html
@@ -1,0 +1,15 @@
+<zeppelin-notebook-paragraph-dynamic-forms
+  *ngIf="paragraph"
+  [disable]="paragraph.status == 'RUNNING' || paragraph.status == 'PENDING'"
+  [paramDefs]="paragraph.settings.params"
+  [formDefs]="paragraph.settings.forms"
+  [runOnChange]="paragraph.config.runOnSelectionChange"
+  (formChange)="runParagraph()">
+</zeppelin-notebook-paragraph-dynamic-forms>
+<zeppelin-notebook-paragraph-result *ngFor="let result of results; index as i; trackBy: trackByIndexFn"
+                                    [id]="paragraph.id"
+                                    [published]="true"
+                                    [currentCol]="paragraph.config.colWidth"
+                                    [config]="configs[i]"
+                                    [result]="result">
+</zeppelin-notebook-paragraph-result>
