+++ zeppelin-web-angular/src/app/pages/workspace/share/dynamic-forms/dynamic-forms.component.less
