+++ zeppelin-web-angular/src/app/pages/workspace/share/result/result.component.less
